Rewrote squares.ts. Started by changing how the squares will be index in later maps by using a string index (i.e a0 or p15) which will hopefully be more immediately obvious than the previous x + y*width stuff. Fixed total board width = total board height = 16 (8 normal board, 4 either side) as won't need more than that.
Defined several new types of Hypersquare. Links are meant to be bidirectional inbetween squares that end in either a Hyper, Arch or Circle. Each of the latter square types changes the input move vector in some way that is in keeping with their geometry (which will be defined in board). They have an invert method that will be used in game/logic.ts once a hypersquare is detected to remap move vector properly.

Added some value checks to these constructors to make sure our values make sense and some comments to make things clear.