Rewrite 3:
Point is now a class with add and multiply methods, rest of squares.ts the same
Board is now a sub child of array - can now index it with square brackets.
Loop string changed to be of format 'atL1L2%...' where 'a' is loop alignment, 't' is loop type and L1, L2 are labels representing the loop points. Can be a pair, full loop or single square.
Can add a pair or single square to the board. Full loops yet to be implemented - goal is to create a 'geometry.ts' module that spits out all the squares belonging to a line or arch etc to be modified later. geometry will also output the visual points beloning to a loop.
Added methods to 'game.ts' to find valid pawn and castling moves. Fixed bug in raycast. Added a make move function that does as expect. Has lots of ugly code to handle edge cases like en passant, castling and promotion (which all seem to work).
'graphics.ts' is still functional (and functional!). Added a bunch of types. Will need to refactor to make wholly procedural/functional (i.e less or more global variables). Still need to add code to draw curves (once loop structures added to 'board.ts')
Still need to add the full game rules to game (can only move pieces on your turn; win conditions)