Board.ts:   fixed odd arches by removing duplicate middle square formed by overlapping make_diag()
Game.ts:
  Fixed bugs in the hypersquare move inverting: used to invert whenever square was on link square, now only loops when impiging on an exit hypersquare (i.e end of an arch/circle/line).
  Added some new game boards: the original hyperchess board and cylindrical (lengthways) chess! Super easy to make with a python script!

Graphics.ts
  Lots of 'orrible changes made to make sure arches display properly. Mapped the align of the arch to an orientation of the unit circle and an offset relative to their centre square (odd dividing by 2 behaviour meant shifts for left and top arches).
  Fixed how radii calculated i.e not using floor/ceil as that didn't work for odd length loops.
  Fixed midpoint calculation for arches, wasn't previously converting to radians (duh).
  Added a check when clicking to ignore click matches from a forbidden square (in case there's an overlap with a curve, we want to always prefer the curve).

It works! It really actually kinda works!

Big bug still outstanding: if you start a piece on a hypersquare or link square, it will raycast using the very first move vector it has (say up for a rook), which will jump to the different link squares, add them to valid moves but stop at the hypersquare exit as the move vector is wrong and can't continue. However this means when it comes later to a later move vector of the piece during the same valid_moves() scan, all those hyper/link squares are already in valid moves, so it terminates, even though that move vector would have allowed it to continue past the exit (i.e is the right way around). This is what causes the pieces to get 'stuck' if they end on a hyper/link square, which is not ideal.